requirements.txt
keyring
PyQt5
python-dotenv
